refactor(core): pipeline-worker dispatcher 정리 + normalize 진입 분리
615줄 → 60줄 dispatcher. normalize 진입(자식 결과 수집 + 소스별 위임)을
pipeline-worker-normalize.ts로 추출. createPipelineHandler 시그니처는
Promise<any> → Promise<unknown>로 타입 강화 (BullMQ 호환 유지).
unknown job.name 경고 로그 + 명시적 undefined 반환 추가.

리팩토링 마스터플랜 Phase 2 PR 2-F 마무리.
